# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Absher Travel & Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /offers:
    get:
      operationId: listOffers
      tags: [offers]
      summary: List all offers
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createOffer
      tags: [offers]
      summary: Create an offer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /offers/{id}:
    get:
      operationId: getOffer
      tags: [offers]
      summary: Get an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Offer details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    patch:
      operationId: updateOffer
      tags: [offers]
      summary: Update an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated offer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteOffer
      tags: [offers]
      summary: Delete an offer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /destinations:
    get:
      operationId: listDestinations
      tags: [destinations]
      summary: List all destinations
      responses:
        "200":
          description: List of destinations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Destination"
    post:
      operationId: createDestination
      tags: [destinations]
      summary: Create a destination
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
  /destinations/{id}:
    get:
      operationId: getDestination
      tags: [destinations]
      summary: Get a destination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Destination details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
    patch:
      operationId: updateDestination
      tags: [destinations]
      summary: Update a destination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DestinationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Destination"
    delete:
      operationId: deleteDestination
      tags: [destinations]
      summary: Delete a destination
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List tourist programs
      parameters:
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of programs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Program"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Create a program
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
  /programs/{id}:
    get:
      operationId: getProgram
      tags: [programs]
      summary: Get a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Program details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    patch:
      operationId: updateProgram
      tags: [programs]
      summary: Update a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgramUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Program"
    delete:
      operationId: deleteProgram
      tags: [programs]
      summary: Delete a program
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /flights/search:
    get:
      operationId: searchFlights
      tags: [flights]
      summary: Search for flights across all providers
      parameters:
        - name: origin
          in: query
          schema: { type: string, minLength: 3, maxLength: 3 }
          description: Origin IATA code (e.g. CAI, DXB)
        - name: destination
          in: query
          schema: { type: string, minLength: 3, maxLength: 3 }
          description: Destination IATA code
        - name: departureDate
          in: query
          schema: { type: string, format: date }
          description: Departure date YYYY-MM-DD
        - name: returnDate
          in: query
          schema: { type: string, format: date }
          description: Return date for round trips
        - name: tripType
          in: query
          schema:
            type: string
            enum: [one_way, round_trip, multi_city]
            default: one_way
        - name: cabinClass
          in: query
          schema:
            type: string
            enum: [economy, premium_economy, business, first]
            default: economy
        - name: adults
          in: query
          schema: { type: integer, minimum: 1, maximum: 9, default: 1 }
        - name: children
          in: query
          schema: { type: integer, minimum: 0, maximum: 8, default: 0 }
        - name: infants
          in: query
          schema: { type: integer, minimum: 0, maximum: 4, default: 0 }
        - name: currency
          in: query
          schema: { type: string, default: USD }
        - name: sort
          in: query
          schema:
            type: string
            enum: [cheapest, fastest, best_value]
            default: cheapest
        - name: legs
          in: query
          schema: { type: string }
          description: JSON array of legs for multi-city (overrides origin/destination)
      responses:
        "200":
          description: Flight search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlightSearchResults"
        "400":
          description: Invalid search parameters
...
components:
  schemas:
...
    Offer:
      type: object
      required: [id, titleAr, titleEn, descriptionAr, descriptionEn, price, duration, imageUrl, featured, createdAt]
      properties:
        id: { type: integer }
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        price: { type: number }
        currency: { type: string }
        duration: { type: string }
        imageUrl: { type: string }
        destination: { type: ["string", "null"] }
        featured: { type: boolean }
        createdAt: { type: string }
    OfferInput:
      type: object
      required: [titleAr, titleEn, descriptionAr, descriptionEn, price, duration, imageUrl]
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        price: { type: number }
        currency: { type: string }
        duration: { type: string }
        imageUrl: { type: string }
        destination: { type: string }
        featured: { type: boolean }
    OfferUpdate:
      type: object
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        price: { type: number }
        currency: { type: string }
        duration: { type: string }
        imageUrl: { type: string }
        destination: { type: string }
        featured: { type: boolean }
    Destination:
      type: object
      required: [id, nameAr, nameEn, descriptionAr, descriptionEn, imageUrl, country, createdAt]
      properties:
        id: { type: integer }
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        imageUrl: { type: string }
        country: { type: string }
        popular: { type: boolean }
        createdAt: { type: string }
    DestinationInput:
      type: object
      required: [nameAr, nameEn, descriptionAr, descriptionEn, imageUrl, country]
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        imageUrl: { type: string }
        country: { type: string }
        popular: { type: boolean }
    DestinationUpdate:
      type: object
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        imageUrl: { type: string }
        country: { type: string }
        popular: { type: boolean }
    DailyItineraryItem:
      type: object
      required: [day, titleAr, titleEn, descriptionAr, descriptionEn]
      properties:
        day: { type: integer }
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
    ProgramHotel:
      type: object
      required: [nameAr, nameEn, stars, city]
      properties:
        nameAr: { type: string }
        nameEn: { type: string }
        stars: { type: integer }
        city: { type: string }
    Program:
      type: object
      required: [id, titleAr, titleEn, price, days, imageUrl, status, featured, isActive, createdAt]
      properties:
        id: { type: integer }
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        country: { type: string }
        cities:
          type: array
          items: { type: string }
        price: { type: number }
        currency: { type: string }
        days: { type: integer }
        nights: { type: integer }
        imageUrl: { type: string }
        images:
          type: array
          items: { type: string }
        dailyItinerary:
          type: array
          items: { $ref: "#/components/schemas/DailyItineraryItem" }
        hotels:
          type: array
          items: { $ref: "#/components/schemas/ProgramHotel" }
        airlines:
          type: array
          items: { type: string }
        includedServices:
          type: array
          items: { type: string }
        excludedServices:
          type: array
          items: { type: string }
        included: { type: ["string", "null"] }
        bookingTerms: { type: ["string", "null"] }
        cancellationPolicy: { type: ["string", "null"] }
        destination: { type: ["string", "null"] }
        status:
          type: string
          enum: [active, featured, new, special_offer, expired, hidden]
        featured: { type: boolean }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    ProgramInput:
      type: object
      required: [titleAr, titleEn, price, days]
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        country: { type: string }
        cities:
          type: array
          items: { type: string }
        price: { type: number }
        currency: { type: string }
        days: { type: integer }
        nights: { type: integer }
        imageUrl: { type: string }
        images:
          type: array
          items: { type: string }
        dailyItinerary:
          type: array
          items: { $ref: "#/components/schemas/DailyItineraryItem" }
        hotels:
          type: array
          items: { $ref: "#/components/schemas/ProgramHotel" }
        airlines:
          type: array
          items: { type: string }
        includedServices:
          type: array
          items: { type: string }
        excludedServices:
          type: array
          items: { type: string }
        included: { type: string }
        bookingTerms: { type: string }
        cancellationPolicy: { type: string }
        destination: { type: string }
        status:
          type: string
          enum: [active, featured, new, special_offer, expired, hidden]
        featured: { type: boolean }
        isActive: { type: boolean }
    ProgramUpdate:
      type: object
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        descriptionAr: { type: string }
        descriptionEn: { type: string }
        country: { type: string }
        cities:
          type: array
          items: { type: string }
        price: { type: number }
        currency: { type: string }
        days: { type: integer }
        nights: { type: integer }
        imageUrl: { type: string }
        images:
          type: array
          items: { type: string }
        dailyItinerary:
          type: array
          items: { $ref: "#/components/schemas/DailyItineraryItem" }
        hotels:
          type: array
          items: { $ref: "#/components/schemas/ProgramHotel" }
        airlines:
          type: array
          items: { type: string }
        includedServices:
          type: array
          items: { type: string }
        excludedServices:
          type: array
          items: { type: string }
        included: { type: string }
        bookingTerms: { type: string }
        cancellationPolicy: { type: string }
        destination: { type: string }
        status:
          type: string
          enum: [active, featured, new, special_offer, expired, hidden]
        featured: { type: boolean }
        isActive: { type: boolean }
...
    FlightSegment:
      type: object
      required: [legOrder, segmentOrder, flightNumber, airlineIata, airlineName, originIata, destinationIata, departureAt, arrivalAt, durationMin, cabinClass]
      properties:
        legOrder: { type: integer }
        segmentOrder: { type: integer }
        flightNumber: { type: string }
        airlineIata: { type: string }
        airlineName: { type: string }
        airlineLogoUrl: { type: ["string", "null"] }
        operatingAirlineIata: { type: ["string", "null"] }
        originIata: { type: string }
        originCity: { type: string }
        destinationIata: { type: string }
        destinationCity: { type: string }
        departureAt: { type: string }
        arrivalAt: { type: string }
        durationMin: { type: integer }
        aircraftType: { type: ["string", "null"] }
        cabinClass: { type: string }
    FlightOffer:
      type: object
      required: [providerSlug, providerOfferId, totalPrice, baseFare, taxes, currency, stops, totalDurationMin, isRefundable, baggageIncludedKg, carryOnIncluded, segments]
      properties:
        providerSlug: { type: string }
        providerOfferId: { type: string }
        totalPrice: { type: number }
        baseFare: { type: number }
        taxes: { type: number }
        currency: { type: string }
        stops: { type: integer }
        totalDurationMin: { type: integer }
        isRefundable: { type: boolean }
        baggageIncludedKg: { type: integer }
        carryOnIncluded: { type: boolean }
        deeplinkUrl: { type: ["string", "null"] }
        segments:
          type: array
          items:
            $ref: "#/components/schemas/FlightSegment"
    ProviderSummary:
      type: object
      required: [provider, count, lowestPrice, currency]
      properties:
        provider: { type: string }
        count: { type: integer }
        lowestPrice: { type: number }
        currency: { type: string }
    FlightSearchResults:
      type: object
      required: [searchHash, totalResults, providerSummary, offers]
      properties:
        searchHash: { type: string }
        cachedAt: { type: ["string", "null"] }
        totalResults: { type: integer }
        providerSummary:
          type: array
          items:
            $ref: "#/components/schemas/ProviderSummary"
        offers:
          type: array
          items:
            $ref: "#/components/schemas/FlightOffer"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT